+++ changelog.md
@@ -1,5 +1,7 @@
 # v1.0.3
-- Added developer label to mod list item
+- Added toggle setting `Minimalist List`
+  - Added optional developer label to mod list item
+  - Version label in mod list item will only appear if *`Minimalist List`* is enabled
 - Fixed coloring for version and ID label in mod list item
 - Adjusted positioning for clear search button and search box on pop-up
 - Added guide text for first-time users on *Favorite Mods* mod list item