+++ changelog.md
@@ -4,6 +4,7 @@
   - Version label in mod list item will only appear if *`Minimalist List`* is enabled
 - Fixed coloring for version and ID label in mod list item
 - Fixed enabled-only filter on mod list
+- Adjusted positioning for mod settings button on pop-up
 - Adjusted positioning for clear search button and search box on pop-up
 - Added guide text for first-time users on *Favorite Mods* mod list item
 - Added and tweaked more node IDs